# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "vLLM CI Dashboard API",
    "version": "1.0.0",
    "description": "Public read-only API backing https://ci.vllm.ai. All documented GET endpoints are unauthenticated and return JSON; errors use `{ \"error\": string }`. Responses are cached via `Vercel-CDN-Cache-Control`; treat data as eventually consistent (30s-10min staleness)."
  },
  "servers": [
    {
      "url": "https://ci.vllm.ai",
      "description": "Production"
    }
  ],
  "paths": {
...
    "/builds/trace": {
      "get": {
        "operationId": "get-builds-trace",
        "summary": "OTel span lanes for a single CI build",
        "description": "available=false with empty lanes when the build has no spans. Whole-build views cap at 5000 spans; job detail pages at 2000.",
        "tags": [
          "Builds"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "organization",
            "schema": {
              "type": "string",
              "description": "required, 400 when missing/invalid"
            },
            "required": false,
            "description": "required, 400 when missing/invalid"
          },
          {
            "in": "query",
            "name": "pipeline",
            "schema": {
              "type": "string",
              "description": "required, 400 when missing/invalid"
            },
            "required": false,
            "description": "required, 400 when missing/invalid"
          },
          {
            "in": "query",
            "name": "buildNumber",
            "schema": {
              "type": "string",
              "description": "required, 400 when missing/invalid"
            },
            "required": false,
            "description": "required, 400 when missing/invalid"
          },
          {
            "in": "query",
            "name": "jobId",
            "schema": {
              "type": "string",
              "description": "UUID; switches to paged job detail spans"
            },
            "required": false,
            "description": "UUID; switches to paged job detail spans"
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "number",
              "description": "0-based (default 0); must be 0 unless jobId is set"
            },
            "required": false,
            "description": "0-based (default 0); must be 0 unless jobId is set"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
    "/eval/filters": {
      "get": {
        "operationId": "get-eval-filters",
        "summary": "Available eval filter values",
        "description": "Distinct models, tasks, images, metric filters, and metric names in the given date range.",
        "tags": [
          "Eval"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "start",
            "schema": {
              "type": "string",
              "description": "Any Date.parse-able date (e.g. YYYY-MM-DD); invalid values are ignored"
            },
            "required": false,
            "description": "Any Date.parse-able date (e.g. YYYY-MM-DD); invalid values are ignored"
          },
          {
            "in": "query",
            "name": "end",
            "schema": {
              "type": "string",
              "description": "Any Date.parse-able date; a bare YYYY-MM-DD is bumped to end-of-day so the day is inclusive"
            },
            "required": false,
            "description": "Any Date.parse-able date; a bare YYYY-MM-DD is bumped to end-of-day so the day is inclusive"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
...
  },
...
}